Committing per CR-12 1331
I have edited the file gmOiwfs.pv (located on pisces in directory /gemini/GEM8.6/gmos/gmos/data) so that the new probe mapping I have implemented for GMOS-N will be robust against a CC crash or reboot.  We want these values preserved so that they are not lost the next time the GMOS CC software is upgraded (or has to be restored from the repository) so please commit this file to CVS/SVN.  Thanks!

This update was necessary after prism re-alignment work done May 15, 2012.